# backend/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Audit Wolf API
  description: |
    Comprehensive smart contract auditing platform API that combines AI-powered analysis 
    with traditional static analysis tools to provide detailed security assessments of Solidity contracts.
  version: 1.0.0
  contact:
    name: Audit Wolf Support
    email: support@auditwolf.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://api.auditwolf.com/api
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /contracts:
    get:
      tags:
        - Contracts
      summary: Get user contracts
      description: Retrieve all contracts belonging to the authenticated user
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Contracts retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      tags:
        - Contracts
      summary: Create new contract
      description: Upload or paste a new Solidity contract for analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContractRequest"
      responses:
        "201":
          description: Contract created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractResponse"
        "400":
          description: Invalid contract data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /contracts/{id}:
    get:
      tags:
        - Contracts
      summary: Get contract by ID
      description: Retrieve a specific contract by its ID
      parameters:
        - name: id
          in: path
          required: true
          description: Contract UUID
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Contract retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractResponse"
        "404":
          description: Contract not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      tags:
        - Contracts
      summary: Update contract
      description: Update contract metadata
      parameters:
        - name: id
          in: path
          required: true
          description: Contract UUID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContractRequest"
      responses:
        "200":
          description: Contract updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractResponse"
        "404":
          description: Contract not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags:
        - Contracts
      summary: Delete contract
      description: Delete a contract and all associated audits
      parameters:
        - name: id
          in: path
          required: true
          description: Contract UUID
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Contract deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Contract not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /contracts/validate:
    post:
      tags:
        - Contracts
      summary: Validate contract source code
      description: Validate Solidity source code without creating a contract
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateContractRequest"
      responses:
        "200":
          description: Validation completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationResponse"

  # Analysis Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Error message"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string

    # Health Schemas
...
    CreateContractRequest:
      type: object
      required:
        - name
        - sourceCode
      properties:
        name:
          type: string
          example: "MyToken"
        sourceCode:
          type: string
          example: "pragma solidity ^0.8.0;\n\ncontract MyToken {\n    // Contract code here\n}"
        compilerVersion:
          type: string
          example: "0.8.19"
    UpdateContractRequest:
      type: object
      properties:
        name:
          type: string
          example: "UpdatedTokenName"
    Contract:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        name:
          type: string
        source_code:
          type: string
        compiler_version:
          type: string
        file_hash:
          type: string
        created_at:
          type: string
          format: date-time
    ContractResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: "#/components/schemas/Contract"
    ContractsResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: "#/components/schemas/Contract"
    ValidateContractRequest:
      type: object
      required:
        - sourceCode
      properties:
        sourceCode:
          type: string
          example: "pragma solidity ^0.8.0;\n\ncontract MyToken {\n    // Contract code here\n}"
    ValidationResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            isValid:
              type: boolean
            errors:
              type: array
              items:
                type: string
            metrics:
              type: object
              properties:
                complexity:
                  type: integer
                lines_of_code:
                  type: integer
                functions_count:
                  type: integer

    # Analysis Schemas
...